auto_paster/requirements.txt
Pillow
pytesseract
pyperclip
keyboard
plyer